requirements.txt
streamlit==1.56.0
opencv-python==4.13.0.92
numpy==2.4.4
mediapipe==0.10.33
torch==2.11.0
torchvision==0.26.0
Pillow==12.2.0